requirements.txt
django==2.0.6
djangorestframework==3.8.2
markdown==2.6.8
django-filter==1.1.0
django-cors-headers==2.2.1
gunicorn==19.9.0
social-auth-app-django==2.1.0
